requirements.txt
bidict==0.22.1
bitarray==2.8.1
contourpy==1.1.1
cycler==0.10.0
decorator==4.4.1
fonttools==4.43.0
imageio==2.31.5
kiwisolver==1.1.0
lazy_loader==0.3
matplotlib==3.8.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.1
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.3
six==1.16.0
tifffile==2023.9.26